backend/requirements.txt
fastapi[standard]
pydantic-settings
asyncpg
sqlmodel
alembic
pyjwt
redis
resend
itsdangerous